@@ -6,7 +6,7 @@
     "minimum-stability": "dev",
     "prefer-stable": true,
     "require": {
-        "php": "^7.4 | ^8.1",
+        "php": "^8.1",
         "ext-ctype": "*",
         "ext-iconv": "*",
         "ext-sqlite3": "*",
@@ -16,28 +16,27 @@
         "doctrine/doctrine-migrations-bundle": "^3.0",
         "doctrine/orm": "^2.8",
         "neitanod/forceutf8": "^2.0",
-        "symfony/console": "5.2.*",
-        "symfony/dotenv": "5.2.*",
-        "symfony/filesystem": "5.2.*",
-        "symfony/finder": "5.2.*",
+        "symfony/console": "^6.1",
+        "symfony/dotenv": "^6.1",
+        "symfony/filesystem": "^6.1",
+        "symfony/finder": "^6.1",
         "symfony/flex": "^1.3.1",
-        "symfony/framework-bundle": "5.2.*",
-        "symfony/http-client": "5.2.*",
+        "symfony/framework-bundle": "^6.1",
+        "symfony/http-client": "^6.1",
         "symfony/monolog-bundle": "^3.6",
-        "symfony/proxy-manager-bridge": "5.2.*",
-        "symfony/twig-bundle": "5.2.*",
-        "symfony/yaml": "5.2.*",
+        "symfony/proxy-manager-bridge": "^6.1",
+        "symfony/twig-bundle": "^6.1",
+        "symfony/yaml": "^6.1",
         "twig/twig": "^3.0"
     },
     "require-dev": {
         "escapestudios/symfony2-coding-standard": "^3.11",
-        "friendsofphp/php-cs-fixer": "^2.17",
-        "phpunit/phpunit": "^9.5",
-        "psalm/plugin-symfony": "^2.1",
-        "symfony/browser-kit": "^5.2",
-        "symfony/css-selector": "^5.2",
+        "friendsofphp/php-cs-fixer": "^3.8",
+        "psalm/plugin-symfony": "^3.1",
+        "symfony/browser-kit": "^6.1",
+        "symfony/css-selector": "^6.1",
         "symfony/maker-bundle": "^1.26",
-        "symfony/phpunit-bridge": "^5.3",
+        "symfony/phpunit-bridge": "^6.1",
         "vimeo/psalm": "^4.3"
     },
     "config": {
@@ -95,7 +94,7 @@
     "extra": {
         "symfony": {
             "allow-contrib": false,
-            "require": "5.2.*"
+            "require": "6.1.*"
         }
     }
 }
